+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Added the option `--with-dependencies` / `-w` to the `fixture:run:group` command
+  - This command recursively runs all fixtures that are required by the given fixture even if they are not in the same group
+- Added new method to the fixture trait: `runFixtureGroupWithDependencies()`
+
 ## [2.1.0] - 2022-09-19
 ### Added
 - Added the option `--with-dependencies` / `-w` to the `fixture:run:single` command